requirements.txt
aniso8601==8.0.0
autopep8==1.7.0
azure-core==1.26.0
azure-storage-blob==12.13.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
cryptography==38.0.1
dnspython==2.2.1
email-validator==1.3.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-WTF==1.0.1
idna==3.4
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==2.2.0
MarkupSafe==1.1.1
msrest==0.7.1
numpy==1.23.4
oauthlib==3.2.1
opencv-python==4.6.0.66
ptyprocess==0.7.0
pycodestyle==2.9.1
pycparser==2.21
pytz==2020.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==1.0.1
WTForms==3.0.1
